requirements.txt
aiofiles
asyncio
customtkinter 
darkdetect 
greenlet 
macholib 
numpy 
packaging 
pandas 
pillow 
playwright 
psycopg2-binary 
pyee 
pyinstaller 
pyinstaller-hooks-contrib 
python-dotenv 
watchdog 
nest-asyncio
plyer
pyobjus